Re-work the main layout to be based on my 2-column map viewer example template (http://jumpinjackie.github.io/bootstrap-viewer-template/2-column/index.html)
The benefit of this layout is that we don't have a schizophrenic UI for phablet/mini-tablet displays. As long as the hamburger icon is visible in the navbar, the rest of the layout will act as though it were a phone-based display and the sidebar will fill the entire screen.